Merged Files_Added Search Func_Added Menu
Used Shawn's merged files and added menu.c and menu.h so that the user can select to search for a word or print out all of the words and their count. Added the search function needed for the BST. Code has been tested with CLion and csegrid and appears to be working. When compiling on the csegrid must use gcc -std=c11 main.c menu.c etc. etc.