Modified writeFile function
Slightly adjusted Sara's code to so that argv[1] is parsed to use number from inputXX.txt to write output file as myoutputXX.txt as per instructor requirements listed on Slack. Also created a separate recursive tree write function in outputFile.c to minimize the chance for error with the recursive calls. Tested on both CLion and CSEGrid and works. Input validation not added for the file names. Will do if I enough time.